requirements.txt
thaiaddress
starlette==0.13.6
fastapi
uvicorn
pymessenger
python-decouple